+++ bundle/src/main/java/com/adobe/acs/commons/wcm/impl/PageRootProviderMultiImpl.java
@@ -21,13 +21,10 @@ import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.felix.scr.annotations.Activate;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.Deactivate;
-import org.apache.felix.scr.annotations.Reference;
-import org.apache.felix.scr.annotations.ReferenceCardinality;
-import org.apache.felix.scr.annotations.ReferencePolicy;
-import org.apache.felix.scr.annotations.Service;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.commons.osgi.Order;
 import org.apache.sling.commons.osgi.RankedServices;
@@ -38,20 +35,21 @@ import com.adobe.acs.commons.wcm.PageRootProvider;
 import com.day.cq.wcm.api.Page;
 import com.day.cq.wcm.api.PageManager;
 
-@Component(metatype=false)
-@Service(PageRootProvider.class)
 /**
  * Service to fetch the site root page (i.e. home page) for a given resource.
  * Supports multiple (independent) configurations.
  *
  * @see PageRootProviderConfig
  */
+@Component(service = PageRootProvider.class)
 public class PageRootProviderMultiImpl implements PageRootProvider {
 
-    private static final Logger log = LoggerFactory.getLogger(PageRootProviderMultiImpl.class);
+    private static final Pattern VERSION_HISTORY_PATTERN = Pattern.compile("/tmp/versionhistory/[0-9a-f]+/[0-9a-f-]+/(.*)");
+    private static final Pattern XF_PATH_PATTERN = Pattern.compile("/content/experience-fragments/(.*)");
+    private static final Pattern LAUNCH_PATH_PATTERN = Pattern.compile("/content/launches/.*?/content/(.*)");
+    private static final Logger LOG = LoggerFactory.getLogger(PageRootProviderMultiImpl.class);
 
-    @Reference(name = "config", referenceInterface = PageRootProviderConfig.class, cardinality = ReferenceCardinality.MANDATORY_MULTIPLE, policy = ReferencePolicy.DYNAMIC)
-    private RankedServices<PageRootProviderConfig> configList = new RankedServices<>(Order.ASCENDING);
+    private final RankedServices<PageRootProviderConfig> configList = new RankedServices<>(Order.ASCENDING);
 
     @Override
     public Page getRootPage(Resource resource) {
@@ -62,9 +60,9 @@ public class PageRootProviderMultiImpl implements PageRootProvider {
             Page rootPage = pageManager.getPage(pagePath);
 
             if (rootPage == null) {
-                log.debug("Page Root not found at [ {} ]", pagePath);
+                LOG.debug("Page Root not found at [ {} ]", pagePath);
             } else if (!rootPage.isValid()) {
-                log.debug("Page Root invalid at [ {} ]", pagePath);
+                LOG.debug("Page Root invalid at [ {} ]", pagePath);
             } else {
                 return rootPage;
             }
@@ -76,21 +74,39 @@ public class PageRootProviderMultiImpl implements PageRootProvider {
     @Override
     public String getRootPagePath(String resourcePath) {
         for (PageRootProviderConfig config : this.configList) {
+            String pathToSearch = resourcePath;
+
+            // If page/XF history viewer should use the corresponding live content tree to determine the root...
+            if (config.getHistoryViewerFallback()) {
+                pathToSearch = VERSION_HISTORY_PATTERN.matcher(pathToSearch).replaceFirst("/content/$1");
+            }
+
+            // If launch content should use the corresponding live content tree to determine the root...
+            if (config.getLaunchFallback()) {
+                pathToSearch = LAUNCH_PATH_PATTERN.matcher(pathToSearch).replaceFirst("/content/$1");
+            }
+
+            // If XF should use the corresponding site content tree to determine the root...
+            if ("site".equals(config.getXfRootPathMethod())) {
+                pathToSearch = XF_PATH_PATTERN.matcher(pathToSearch).replaceFirst("/content/$1");
+            }
+
             for (Pattern pattern : config.getPageRootPatterns()) {
-                final Matcher matcher = pattern.matcher(resourcePath);
+                final Matcher matcher = pattern.matcher(pathToSearch);
 
                 if (matcher.find()) {
                     String rootPath = matcher.group(1);
-                    log.debug("Page Root found at [ {} ]", rootPath);
+                    LOG.debug("Page Root found at [ {} ]", rootPath);
                     return rootPath;
                 }
             }
         }
 
-        log.debug("Resource path does not include the configured page root path.");
+        LOG.debug("Resource path does not include the configured page root path.");
         return null;
     }
 
+    @Reference(name = "config", service = PageRootProviderConfig.class, cardinality = ReferenceCardinality.AT_LEAST_ONE, policy = ReferencePolicy.DYNAMIC)
     protected void bindConfig(final PageRootProviderConfig config, Map<String, Object> props) {
         this.configList.bind(config, props);
     }